First commit
Made some changes and focused on the reset password functionality
