+++ src/Server/controllers/authController.js
@@ -1,6 +1,5 @@
 const { getUserByEmail, createUser } = require('../models/User');
 const { hashPassword, comparePassword } = require('../utils/passwordUtils');
-const { verifyToken } = require('../utils/authUtils');
 const jwt = require('jsonwebtoken');
 
 require('dotenv').config({ path: './.env' });
@@ -27,7 +26,6 @@ async function signup(req, res) {
       house,
       landmark
     });
-
     res.status(201).json({ message: 'User created successfully', userId: newUser.insertedId });
   } catch (error) {
     res.status(500).json({ message: 'Error creating user', error: error.message });
@@ -38,60 +36,35 @@ async function login(req, res) {
   const { email, password } = req.body;
 
   try {
-      const user = await getUserByEmail(email);
-      if (!user || !(await comparePassword(password, user.password))) {
-          return res.status(400).json({ message: 'Invalid credentials' });
-      }
+    const user = await getUserByEmail(email);
+    console.log(user.email)
+    if (!user || !(await comparePassword(password, user.password))) {
+      return res.status(400).json({ message: 'Invalid credentials' });
+    }
 
-      // Create a JWT token with 30 days expiration
-      const token = jwt.sign(
-          { userId: user._id },
-          JWT_SECRET,
-          { expiresIn: '30d' } // 30 days expiry
-      );
+    // Create a JWT token with 30 days expiration
+    const token = jwt.sign(
+      { userId: user._id },
+      JWT_SECRET,
+      { expiresIn: '30d' } // 30 days expiry
+    );
 
-      // Send the token and user info back
-      res.status(200).json({ message: 'Login successful', token, userId: user._id });
+    // Send the token and user info back
+    res.status(200).json({ message: 'Login successful', token, userId: user._id });
   } catch (error) {
-      res.status(500).json({ message: 'Error logging in', error: error.message });
+    res.status(500).json({ message: 'Error logging in', error: error.message });
   }
 }
 
-async function verifyUserToken(req, res) {
+async function resetPassword(req, res) {
+  // const { email } = req.body;
   try {
-      console.log('Headers received:', req.headers);
-      const authHeader = req.headers.authorization;
-      
-      if (!authHeader) {
-          console.log('No authorization header found');
-          return res.status(401).json({ message: 'No token provided' });
-      }
-
-      console.log('Auth header:', authHeader);
-      const token = authHeader.split(' ')[1];
-      
-      if (!token) {
-          console.log('No token found in auth header');
-          return res.status(401).json({ message: 'No token provided in correct format' });
-      }
-
-      const verification = verifyToken(token);
-      console.log('Verification result:', verification);
-
-      if (!verification.valid) {
-          return res.status(401).json({ message: verification.error || 'Invalid token' });
-      }
-
-      return res.status(200).json({ valid: true, userId: verification.userId });
-
+    const user = await getUserByEmail(email);
+    console.log(user);
+    res.status(200).json({ message: 'user is valid', user: user.name })
   } catch (error) {
-      console.error('Server error in verifyUserToken:', error);
-      return res.status(500).json({ 
-          message: 'Error verifying token', 
-          error: error.message,
-          stack: process.env.NODE_ENV === 'development' ? error.stack : undefined
-      });
+    res.status(500).json({ message: 'User doesn\'t exist', error: error.message });
   }
 }
 
-module.exports = { signup, login, verifyUserToken };
+module.exports = { signup, login, resetPassword };
