+++ src/routes/auth.js
@@ -12,6 +12,11 @@ router.get("/Register", (req, res) => {
   res.sendFile("templates/Auth/signup.html", { root: path.join(__dirname, "../") });
 });
 
+// reset password page
+router.get("/Reset/Password", (req, res) => {
+  res.sendFile("templates/Auth/Reset-Password.html", { root: path.join(__dirname, "../") });
+})
+
 // Submit email page
 router.get("/Reset/Submit/Email", (req, res) => {
   res.sendFile("templates/Auth/Submit-Email.html", { root: path.join(__dirname, "../") });
@@ -20,11 +25,6 @@ router.get("/Reset/Submit/Email", (req, res) => {
 // Submit OTP page
 router.get("/Reset/Submit/OTP", (req, res) => {
   res.sendFile("templates/Auth/SubmitOTP.html", { root: path.join(__dirname, "../") });
-})
-
-// reset password page
-router.get("/Reset/Password", (req, res) => {
-  res.sendFile("templates/Auth/Reset-Password.html", { root: path.join(__dirname, "../") });
 })
 
 module.exports = router;
